+++ src/components/card/index.d.ts
@@ -1,0 +1,1 @@
+export { EuiCardGraphic } from './card_graphic';